Update set-up-local-development.rst
This page is already Drupal specific (which is bad), but it doesn't give enough detail to people who need to do local Drupal development with Drush Make (also bad). So I've noted two of the Drupal specific pieces of information and elaborated on the symlinks so that people understand local development better.