Remove mention of restrict_robots
The restrict_robots setting does not have any impact on the x-robots-tag header on non-production environments by design. The setting also does not have any impact on the x-robots-tag header in production environments currently (which is the result of the bug). Removing references to restrict_robots in the public documentation in the meantime.