Ensure failed curl requests do not download code
Adds the --fail (-f) curl option to commands that download code, meaning that
only an error message will be shown and no data will be output. This protects
against the case where incorrect code may be output and executed even if the
HTTP request fails. Nevertheless, the server and TLS connection must still be
trusted.