remove MASTER_USE_GTID from start command example
With the setup provide by this guide, the "slave_pos" is an empty string when starting the replication database. This causes an error in the reading of the bin logs, as it can not find the '' location in the log. Removing this line from the slave definition allows the database to start and copy as expected.